Changed m_frameCount from uint32_t to uint64_t

+++ Gems/Atom/Feature/Common/Code/Source/RayTracing/RayTracingAccelerationStructurePass.h
@@ -50,7 +50,7 @@ namespace AZ
             uint32_t m_rayTracingRevision = 0;
 
             // keeps track of the current frame to determine updates or rebuilds of the skinned BLASes
-            uint32_t m_frameCount = 0;
+            uint64_t m_frameCount = 0;
 
             // the bits of this constant are used to check if a skinned BLAS is going to be rebuilt in any given frame
             static constexpr uint32_t SKINNED_BLAS_REBUILD_FRAME_INTERVAL = 8;
